+++ topobank/analysis/serializers.py
@@ -47,8 +47,9 @@ class AnalysisFunctionSerializer(serializers.HyperlinkedModelSerializer):
 class AnalysisSubjectSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
         model = AnalysisSubject
-        fields = ['id', 'topography', 'surface', 'collection']
+        fields = ['id', 'tag', 'topography', 'surface']
 
+    tag = serializers.HyperlinkedRelatedField(view_name='manager:tag-api-detail', read_only=True)
     topography = serializers.HyperlinkedRelatedField(view_name='manager:topography-api-detail', read_only=True)
     surface = serializers.HyperlinkedRelatedField(view_name='manager:surface-api-detail', read_only=True)