+++ topobank/analysis/urls.py
@@ -18,6 +18,14 @@ urlpatterns += [
     # GET
     # * Triggers analyses if not yet running
     # * Return state of analyses
+    path(
+        'api/result/',
+        view=views.query_analyses,
+        name='status'
+    ),
+    # GET
+    # * Triggers analyses if not yet running
+    # * Return state of analyses
     # * Return plot configuration for finished analyses
     # This is a post request because the request parameters are complex.
     path(