+++ topobank/manager/tests/utils.py
@@ -13,7 +13,7 @@ from django.core.management import call_command
 from django.shortcuts import reverse
 
 from ...users.tests.factories import UserFactory
-from ..models import Surface, SurfaceCollection, TagModel, Topography
+from ..models import Surface, Tag, Topography
 
 FIXTURE_DIR = os.path.join(
     os.path.dirname(os.path.realpath(__file__)),
@@ -83,13 +83,13 @@ class SurfaceFactory(factory.django.DjangoModelFactory):
     creator = factory.SubFactory(UserFactory)
 
 
-class SurfaceCollectionFactory(factory.django.DjangoModelFactory):
-    """Generates a SurfaceCollection."""
+class TagFactory(factory.django.DjangoModelFactory):
+    """Generates a Tag."""
 
     class Meta:
-        model = SurfaceCollection
+        model = Tag
 
-    name = factory.Sequence(lambda n: "surfacecollection-{:05d}".format(n))
+    name = factory.Sequence(lambda n: "tag-{:05d}".format(n))
 
     @factory.post_generation
     def surfaces(self, create, extracted, **kwargs):
@@ -99,7 +99,7 @@ class SurfaceCollectionFactory(factory.django.DjangoModelFactory):
         if extracted:
             # A list of surfaces were passed in, use them for the manytomany field
             for surface in extracted:
-                self.surfaces.add(surface)
+                self.surface_set.add(surface)
 
 
 class Topography1DFactory(factory.django.DjangoModelFactory):
@@ -144,7 +144,7 @@ class TagModelFactory(factory.django.DjangoModelFactory):
     """Generates a tag."""
 
     class Meta:
-        model = TagModel
+        model = Tag
 
     name = factory.Sequence(lambda n: "tag-{}".format(n))